+++ engines/ags/shared/game/room_file.cpp
@@ -439,7 +439,7 @@ HRoomFileError ReadCompSc3Block(RoomStruct *room, Stream *in, RoomFileVersion da
 
 // Room object names
 HRoomFileError ReadObjNamesBlock(RoomStruct *room, Stream *in, RoomFileVersion data_ver) {
-	int name_count = in->ReadByte();
+	size_t name_count = in->ReadByte();
 	if (name_count != room->ObjectCount)
 		return new RoomFileError(kRoomFileErr_InconsistentData,
 			String::FromFormat("In the object names block, expected name count: %d, got %d", room->ObjectCount, name_count));
@@ -455,7 +455,7 @@ HRoomFileError ReadObjNamesBlock(RoomStruct *room, Stream *in, RoomFileVersion d
 
 // Room object script names
 HRoomFileError ReadObjScNamesBlock(RoomStruct *room, Stream *in, RoomFileVersion data_ver) {
-	int name_count = in->ReadByte();
+	size_t name_count = in->ReadByte();
 	if (name_count != room->ObjectCount)
 		return new RoomFileError(kRoomFileErr_InconsistentData,
 			String::FromFormat("In the object script names block, expected name count: %d, got %d", room->ObjectCount, name_count));
